Tweak styling a bunch
The table of contents was simplified; it no longer decorates its
entries with index numbers etc.

The table of contents now also indicates which topics are within the
viewport. At the moment it shows all such topics, though I'd slightly
prefer to show only the most important one (i.e. the one most likely
being viewed by the user).

<dokki-header> was merged into <dokki-side-panel>, which also requires
this change to be made in end-user HTML (see the update to the sample
document).

This is work in progress. Certain features (like the color theme
selector) were removed temporarily, and will be re-integrated in the
near future.