dokki-image: Add support for lazy loading of images
When <dokki-image> is given a 'thumbnail-src' attribute, the attribute's
contents will be set as the image's 'src' attribute until the image
element becomes visible in the viewport, after which the original 'src'
attribute's value will be used.

You'd generally use this feature together with the build-final.js script,
which automatically generates a thumbnail for any <dokki-image> element
whose 'src' attribute points to a locally-hosted image.